+++ ClassTrack/Presentation/ClassTrack.MVC/ViewModels/Quiz/PutQuizVM.cs
@@ -6,7 +6,7 @@ namespace ClassTrack.MVC.ViewModels
 
        string Name,
        DateTime StartTime,
-       int Duration,
+       double Duration,
        long ClassRoomId);
     
 }
